requirements.txt
opencv-python
pyocr
numpy
pytesseract
tqdm
Pillow
moviepy
torch
torchvision